feat: created category specific directories to reorganize files across React components, Redux reducers, and Express routers
