+++ client/components/index.js
@@ -5,8 +5,8 @@
  * having to figure out which file they belong to!
  **/
 
-export {default as Navbar} from './Navbar'
-export {default as Hello} from './Hello'
-export {default as Home} from './Home'
-export {default as PageNotFound} from './PageNotFound'
-export {Login, Signup} from './AuthForm'
+export {default as Navbar} from './navbar/Navbar'
+export {default as Hello} from './navbar/Hello'
+export {default as Home} from './dashboard/Home'
+export {default as PageNotFound} from './404/PageNotFound'
+export {Login, Signup} from './auth/AuthForm'
