+++ server/api/index.js
@@ -2,7 +2,7 @@
 const router = require('express').Router()
 
 // Middleware
-router.use('/users', require('./users'))
+router.use('/users', require('./routers/users'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
