+++ server/api/routers/users.spec.js
@@ -3,8 +3,8 @@
 const {expect} = require('chai')
 const request = require('supertest')
 
-const app = require('../index')
-const db = require('../db')
+const app = require('../../index')
+const db = require('../../db')
 
 const User = db.model('user')
